[edit] 컴포넌트 폴더 경로 및 컴포넌트 명 수정
