+++ src/app/account/mybook/[bookId]/page.tsx
@@ -1,5 +1,5 @@
 import Header from "@/components/home/Header";
-import ReservationItemDetail from "@/components/account/ReservationItemDetail";
+import ReservationDetail from "@/components/account/ReservationDetail";
 interface PageParams {
   params: {
     bookId: string;
@@ -15,7 +15,7 @@ export default function Page({ params }: PageParams) {
       <main
         className={`relative m-auto w-full pt-5 px-[30px] sm:max-fit md:px-[140px] max-w-[360px] sm:max-w-[1000px]`}
       >
-        <ReservationItemDetail bookId={bookId} />
+        <ReservationDetail bookId={bookId} />
       </main>
     </>
   );
