+++ src/components/account/MyReservation.tsx
@@ -1,0 +1,20 @@
+"use client";
+import { useEffect } from "react";
+import useBookingDetail from "@/hooks/useBookingDetail";
+import ReservationList from "./ReservationList";
+
+export default function MyReservation() {
+  const { bookingData, completedBookingData, isLoading, fetchAllBookings } =
+    useBookingDetail();
+
+  useEffect(() => {
+    fetchAllBookings();
+  }, [fetchAllBookings]);
+
+  return (
+    <div className="flex flex-col relative">
+      <ReservationList data={bookingData} />
+      <ReservationList data={completedBookingData} isPastData />
+    </div>
+  );
+}
